+++ UnityProject/Assets/Scripts/Objects/Closets/AbandonedCrates.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 26312ad49a2f46988a77c9630730a98a
+timeCreated: 1661144187
